Created modular functionallity for movies and weather data queries and updated readme for todays assignment
